# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/notifications:
    get:
      operationId: listAdminNotifications
      tags: [applications]
      summary: List committee notifications
      security:
        - committeeSession: []
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: number, minimum: 1, maximum: 100, default: 30 }
      responses:
        "200":
          description: Committee notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/notifications/read-all:
    post:
      operationId: markAllAdminNotificationsRead
      tags: [applications]
      summary: Mark all visible notifications as read
      security:
        - committeeSession: []
      responses:
        "200":
          description: Notifications marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationReadResponse"
  /admin/notifications/{id}/read:
    post:
      operationId: markAdminNotificationRead
      tags: [applications]
      summary: Mark a notification as read
      security:
        - committeeSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationReadResponse"
  /admin/observations:
    get:
      operationId: getAdminObservations
      tags: [applications]
      summary: Get observation aggregates by level
      security:
        - committeeSession: []
      parameters:
        - { name: jenjang, in: query, required: false, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string } }
        - { name: q, in: query, required: false, schema: { type: string } }
        - { name: from, in: query, required: false, schema: { type: string, format: date } }
        - { name: to, in: query, required: false, schema: { type: string, format: date } }
      responses:
        "200":
          description: Observation aggregates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObservationResponse"
  /admin/master-data:
    get:
      operationId: listAdminMasterData
      tags: [applications]
      summary: List master student data
      security:
        - committeeSession: []
      parameters:
        - { name: jenjang, in: query, required: false, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string } }
        - { name: q, in: query, required: false, schema: { type: string } }
        - { name: from, in: query, required: false, schema: { type: string, format: date } }
        - { name: to, in: query, required: false, schema: { type: string, format: date } }
      responses:
        "200":
          description: Master data list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MasterDataResponse"
  /admin/export.xlsx:
    get:
      operationId: exportAdminMasterData
      tags: [applications]
      summary: Export master data as XLSX
      security:
        - committeeSession: []
      parameters:
        - { name: jenjang, in: query, required: false, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string } }
        - { name: q, in: query, required: false, schema: { type: string } }
        - { name: from, in: query, required: false, schema: { type: string, format: date } }
        - { name: to, in: query, required: false, schema: { type: string, format: date } }
      responses:
        "200":
          description: XLSX export
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: {}
  /admin/applications/{id}/files.zip:
    get:
      operationId: downloadApplicationFilesZip
      tags: [applications]
      summary: Download one application files as ZIP
      security:
        - committeeSession: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: Application ZIP
          content:
            application/zip: {}
  /admin/files.zip:
    post:
      operationId: downloadBulkApplicationFilesZip
      tags: [applications]
      summary: Download selected application files as ZIP
      security:
        - committeeSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkFilesRequest"
      responses:
        "200":
          description: Bulk application ZIP
          content:
            application/zip: {}
components:
  securitySchemes:
    committeeSession:
      type: apiKey
      in: cookie
      name: spmb_committee_session
  schemas:
...
    StudentApplicationInput:
      type: object
      required:
        - jenjang
        - nama_calon
        - nama_panggilan
        - jenis_kelamin
        - tempat_lahir
        - tanggal_lahir
        - nik_anak
        - alamat_domisili
        - anak_ke
        - jumlah_saudara
        - status_anak
        - agama
        - warga_negara
        - tinggi_badan
        - berat_badan
        - transportasi
        - jarak_sekolah
        - nomor_kk
        - nik_ayah
        - nik_ibu
        - nomor_hp_orangtua
        - email
        - nama_ayah
        - ttl_ayah
        - pendidikan_ayah
        - pekerjaan_ayah
        - penghasilan_ayah
        - instansi_jabatan_ayah
        - nama_ibu
        - ttl_ibu
        - pendidikan_ibu
        - pekerjaan_ibu
        - penghasilan_ibu
        - instansi_jabatan_ibu
        - foto_3x4
        - akte_lahir
        - kartu_keluarga
        - ktp_orangtua
        - bukti_bayar
      properties:
        jenjang: { type: string, enum: [Playgroup, Daycare, TK-A, TK-B, SD, SMP] }
        nama_calon: { type: string, minLength: 2, maxLength: 100 }
        nama_panggilan: { type: string, minLength: 1, maxLength: 100 }
        jenis_kelamin: { type: string, enum: [Laki-laki, Perempuan] }
        tempat_lahir: { type: string, minLength: 1, maxLength: 100 }
        tanggal_lahir:
          type: string
          format: date
          description: "Usia dihitung pada 1 Juli 2027. Minimum: Playgroup 3 tahun, TK-A 4 tahun, TK-B 5 tahun, SD 6 tahun. Daycare dan SMP tidak memiliki batas usia minimum."
        nisn: { type: string, pattern: "^[0-9]{10}$", maxLength: 10, nullable: true }
        nik_anak: { type: string, pattern: "^[0-9]{16}$", minLength: 16, maxLength: 16 }
        alamat_domisili: { type: string, minLength: 1, maxLength: 1000 }
        anak_ke: { type: string, pattern: "^[0-9]+$", description: "Integer from 1 to 20" }
        jumlah_saudara: { type: string, pattern: "^[0-9]+$", description: "Integer from 0 to 50" }
        status_anak: { type: string, minLength: 1, maxLength: 50 }
        agama: { type: string, minLength: 1, maxLength: 50 }
        warga_negara: { type: string, minLength: 1, maxLength: 50 }
        tinggi_badan: { type: string, pattern: "^[0-9]+([.][0-9]+)?$", description: "Height in centimeters" }
        berat_badan: { type: string, pattern: "^[0-9]+([.][0-9]+)?$", description: "Weight in kilograms" }
        riwayat_penyakit: { type: string, maxLength: 1000, nullable: true }
        transportasi: { type: string, minLength: 1, maxLength: 50 }
        jarak_sekolah: { type: string, minLength: 1, maxLength: 100 }
        nama_sekolah_asal: { type: string, minLength: 1, maxLength: 255, nullable: true, description: "Required for SD and SMP; not required for Playgroup, Daycare, TK-A, or TK-B" }
        tahun_lulus: { type: string, pattern: "^[0-9]{4}$", nullable: true, description: "Graduation year, 1900 or later; required for SD and SMP" }
        alamat_sekolah_asal: { type: string, minLength: 1, maxLength: 1000, nullable: true, description: "Required for SD and SMP; not required for Playgroup, Daycare, TK-A, or TK-B" }
        nomor_kk: { type: string, pattern: "^[0-9]{16}$", minLength: 16, maxLength: 16 }
        nik_ayah: { type: string, pattern: "^[0-9]{16}$", minLength: 16, maxLength: 16 }
        nik_ibu: { type: string, pattern: "^[0-9]{16}$", minLength: 16, maxLength: 16, nullable: true }
        nomor_hp_orangtua: { type: string, minLength: 10, maxLength: 16 }
        email: { type: string, pattern: "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$", maxLength: 254 }
        nama_ayah: { type: string, minLength: 1, maxLength: 100 }
        ttl_ayah: { type: string, minLength: 1, maxLength: 150 }
        pendidikan_ayah: { type: string, minLength: 1, maxLength: 100 }
        pekerjaan_ayah: { type: string, minLength: 1, maxLength: 100 }
        penghasilan_ayah: { type: string, minLength: 1, maxLength: 100 }
        instansi_jabatan_ayah: { type: string, minLength: 1, maxLength: 150 }
        nama_ibu: { type: string, minLength: 1, maxLength: 100 }
        ttl_ibu: { type: string, minLength: 1, maxLength: 150 }
        pendidikan_ibu: { type: string, minLength: 1, maxLength: 100 }
        pekerjaan_ibu: { type: string, minLength: 1, maxLength: 100 }
        penghasilan_ibu: { type: string, minLength: 1, maxLength: 100 }
        instansi_jabatan_ibu: { type: string, minLength: 1, maxLength: 150 }
        nama_wali: { type: string, maxLength: 100, nullable: true }
        hubungan_wali: { type: string, maxLength: 100, nullable: true }
        foto_3x4: { type: string, description: "Uploaded JPG or PNG file, maximum 5 MB" }
        akte_lahir: { type: string, description: "Uploaded PDF, JPG, or PNG file, maximum 5 MB" }
        kartu_keluarga: { type: string, description: "Uploaded PDF, JPG, or PNG file, maximum 5 MB" }
        ktp_orangtua: { type: string, description: "Uploaded PDF, JPG, or PNG file, maximum 5 MB" }
        bukti_bayar: { type: string, description: "Uploaded PDF, JPG, or PNG file, maximum 5 MB" }
...
    ApplicationListItem:
      type: object
      properties:
        id: { type: number }
        nis: { type: string, nullable: true }
        nama_calon: { type: string }
        jenjang: { type: string }
        nama_sekolah_asal: { type: string, nullable: true }
        email: { type: string }
        status: { type: string }
        created_at: { type: string }
      required: [id, nis, nama_calon, jenjang, nama_sekolah_asal, email, status, created_at]
...
    ApplicationFile:
      type: object
      properties:
        field: { type: string }
        label: { type: string }
        url: { type: string }
        available: { type: boolean }
      required: [field, label, url, available]
    ApplicationDetail:
      allOf:
        - $ref: "#/components/schemas/ApplicationListItem"
        - $ref: "#/components/schemas/StudentApplicationInput"
        - type: object
          properties:
            files:
              type: array
              items:
                $ref: "#/components/schemas/ApplicationFile"
          required: [files]
    Notification:
      type: object
      properties:
        id: { type: number }
        application_id: { type: number, nullable: true }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        jenjang: { type: string }
        created_at: { type: string }
        read: { type: boolean }
        nama_calon: { type: string, nullable: true }
        nis: { type: string, nullable: true }
      required: [id, application_id, type, title, message, jenjang, created_at, read, nama_calon, nis]
    NotificationListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Notification" }
        unreadCount: { type: number }
      required: [items, unreadCount]
    NotificationReadResponse:
      type: object
      properties:
        success: { type: boolean }
        count: { type: number, nullable: true }
      required: [success]
    ObservationResponse:
      type: object
      properties:
        jenjang: { type: string }
        total: { type: number }
        counts: { type: object, additionalProperties: { type: number } }
        incompleteDocuments: { type: number }
        trends:
          type: array
          items:
            type: object
            properties:
              date: { type: string }
              count: { type: number }
            required: [date, count]
        breakdowns: { type: object, additionalProperties: true }
      required: [jenjang, total, counts, incompleteDocuments, trends, breakdowns]
    MasterDataResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ApplicationDetail" }
        total: { type: number }
      required: [items, total]
    BulkFilesRequest:
      type: object
      properties:
        ids:
          type: array
          items: { type: number }
          maxItems: 100
      required: [ids]
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        fields:
          type: array
          items: { type: string }
      required: [error]