# lib/api-zod/src/generated/types/masterDataResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationDetail } from './applicationDetail';

export interface MasterDataResponse {
  items: ApplicationDetail[];
  total: number;
}