# lib/api-zod/src/generated/types/notification.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Notification {
  id: number;
  application_id: number | null;
  type: string;
  title: string;
  message: string;
  jenjang: string;
  created_at: string;
  read: boolean;
  nama_calon: string | null;
  nis: string | null;
}