# lib/api-zod/src/generated/types/observationResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ObservationResponseBreakdowns } from './observationResponseBreakdowns';
import type { ObservationResponseCounts } from './observationResponseCounts';
import type { ObservationResponseTrendsItem } from './observationResponseTrendsItem';

export interface ObservationResponse {
  jenjang: string;
  total: number;
  counts: ObservationResponseCounts;
  incompleteDocuments: number;
  trends: ObservationResponseTrendsItem[];
  breakdowns: ObservationResponseBreakdowns;
}